add some system specs
