+++ test/system/static_pages_test.rb
@@ -1,7 +1,7 @@
 require 'application_system_test_case'
 
 class StaticPagesTest < ApplicationSystemTestCase
-  test 'visiting the homepage' do
+  test 'static_pages#landing_page' do
     visit root_url
 
     assert_text 'Connect your Instagram'
